# src/backend/KamraApp.Api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "KamraApp.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/pantry-items": {
      "get": {
        "tags": [
          "PantryItems"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expiringSoon",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PantryItemResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PantryItems"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AddPantryItemRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AddPantryItemRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AddPantryItemRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PantryItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PantryItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PantryItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pantry-items/{id}": {
      "put": {
        "tags": [
          "PantryItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UpdatePantryItemRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UpdatePantryItemRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UpdatePantryItemRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PantryItemResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddPantryItemRequest": {
        "required": [
          "ingredientId",
          "amount",
          "unit"
        ],
        "type": "object",
        "properties": {
          "ingredientId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "unit": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
...
      "Category": {
        "type": "integer"
      },
...
      "PantryItemResponse": {
        "required": [
          "id",
          "ingredientId",
          "ingredientName",
          "amount",
          "unit",
          "quantity",
          "category",
          "expiryDate",
          "expiryEstimated",
          "expired",
          "soonExpiring",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ingredientId": {
            "type": "string",
            "format": "uuid"
          },
          "ingredientName": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "expiryDate": {
            "type": "string",
            "format": "date"
          },
          "expiryEstimated": {
            "type": "boolean"
          },
          "expired": {
            "type": "boolean"
          },
          "soonExpiring": {
            "type": "boolean"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "Unit": {
        "type": "integer"
      },
      "UpdatePantryItemRequest": {
        "required": [
          "amount",
          "unit",
          "category",
          "version"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "unit": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      }
    }
  },
...
}